+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 `RxUtils` adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [1.3.0](https://github.com/APUtils/RxUtils/releases/tag/1.3.0)
+Released on 07/15/2019.
+
+#### Added
+- Added convenience computed properties for CompareResult
+
+#### Changed
+- CompareResult now also returns previous value for same case
+
+
 ## [1.2.0](https://github.com/APUtils/RxUtils/releases/tag/1.2.0)
 Released on 07/15/2019.